Added multi-component symmetries, cosmetic changes only
Main change: added support for multiple symmetric "roots" in the asymmetric unit. For example, modeling a protein "cube" with a fixed trimeric interface at the corners, a fixed dimeric interface at the edges, and a fixed tetrameric interface on the faces. Fun fact: I wrote something similar 3 years ago and never checked it in... in the mean time, symmetry got more complicated... it was harder this time. Check in your shit, then updates are somebody else's problem!

The addition to the symmetry_definition format is very simple: one may now have multiple sets of SUBUNIT jumps with and additional chain and anchor specified for each. Ex:
connect_virtual Jump_to_CompRootA  ROOT       CompRootA
connect_virtual Jump_to_CompCntrA  CompRootA  CompCntrA
connect_virtual Jump_to_CompSubuA1 CompCntrA  CompSubuA1
connect_virtual Jump_to_CompSubuA2 CompCntrA  CompSubuA2
connect_virtual Jump_to_SUBUNIT_A1 CompSubuA1 SUBUNIT A COM
connect_virtual Jump_to_SUBUNIT_A2 CompSubuA2 SUBUNIT A COM
connect_virtual Jump_to_CompRootC  ROOT       CompRootC
connect_virtual Jump_to_CompCntrC  CompRootC  CompCntrC
connect_virtual Jump_to_CompSubuC1 CompCntrC  CompSubuC1
connect_virtual Jump_to_CompSubuC2 CompCntrC  CompSubuC2
connect_virtual Jump_to_SUBUNIT_C1 CompSubuC1 SUBUNIT C 47
connect_virtual Jump_to_SUBUNIT_C2 CompSubuC2 SUBUNIT C 47
connect_virtual Jump_to_CompRootB  ROOT       CompRootB
connect_virtual Jump_to_CompCntrB  CompRootB  CompCntrB
connect_virtual Jump_to_CompSubuB1 CompCntrB  CompSubuB1
connect_virtual Jump_to_CompSubuB2 CompCntrB  CompSubuB2
connect_virtual Jump_to_SUBUNIT_B1 CompSubuB1 SUBUNIT B FIRST_RESIDUE
connect_virtual Jump_to_SUBUNIT_B2 CompSubuB2 SUBUNIT B FIRST_RESIDUE

Also made a minor change to the way SymmData scores the anchor in order so support some extra options, like KEEP_FOLDTREE_ANCHOR, which I sometimes need. This should have only a cosmetic effect.

Added a (text) FoldTree visualization in core/kinematics/util.hh not specific to symmetry; it can be called with only a FoldTree, or with various annotation maps to label nodes and edges. For example, this gnarly 7-component symmetry tree:

=================== SYM FOLD TREE, jump notation: =symfixed= *indep* #symdof# jump[=follows] ========================
ROOT(139)
|----=j36=--->CompRootG(140)----#j29#--->CompCntrG(141)
|                                        |----=j43=--->CompSubuG1(142)----*j15*---->28:Sub1G(28-32)28-----*j6*----->27:Sub1G(24-27)
|                                        \----=j44=--->CompSubuG2(143)---j16=15---->97:Sub2G(97-101)97-----j13=6---->96:Sub2G(93-96)
|----=j37=--->CompRootH(144)----#j30#--->CompCntrH(145)
|                                        |----=j45=--->CompSubuH1(146)----*j17*---->36:Sub1H(33-39)
|                                        \----=j46=--->CompSubuH2(147)---j18=17--->105:Sub2H(102-108)
|----=j38=--->CompRootL(152)----#j31#--->CompCntrL(153)
|                                        |----=j47=--->CompSubuL1(154)----*j19*---->46:Sub1L(44-53)44-----*j7*----->43:Sub1L(40-43)
|                                        \----=j48=--->CompSubuL2(155)---j20=19--->115:Sub2L(113-122)113----j14=7--->112:Sub2L(109-112)
|----=j39=--->CompRootM(148)----#j32#--->CompCntrM(149)
|                                        |----=j49=--->CompSubuM1(150)----*j21*---->59:Sub1M(54-63)
|                                        \----=j50=--->CompSubuM2(151)---j22=21--->128:Sub2M(123-132)
|----=j40=--->CompRootP(156)----#j33#--->CompCntrP(157)
|                                        |----=j51=--->CompSubuP1(158)----*j23*----->6:Sub1P(4-9)
|                                        |                                             |4------*j1*------>3:Sub1P(1-3)
|                                        |                                             \9------*j2*---->Sub1P(10)----*j3*---->Sub1P(11)
|                                        \----=j52=--->CompSubuP2(159)---j24=23---->75:Sub2P(73-78)
|                                                                                      |73-----j8=1----->72:Sub2P(70-72)
|                                                                                      \78-----j9=2---->Sub2P(79)----j10=3--->Sub2P(80)
|----=j41=--->CompRootV(160)----#j34#--->CompCntrV(161)
|                                        |----=j53=--->CompSubuV1(162)----*j25*---->67:Sub1V(64-69)
|                                        \----=j54=--->CompSubuV2(163)---j26=25--->136:Sub2V(133-138)
\----=j42=--->CompRootX(164)----#j35#--->CompCntrX(165)
                                         |----=j55=--->CompSubuX1(166)----*j27*---->18:Sub1X(18-21)
                                         |                                             |18-----*j4*----->17:Sub1X(12-17)
                                         |                                             \21-----*j5*----->22:Sub1X(22-23)
                                         \----=j56=--->CompSubuX2(167)---j28=27---->87:Sub2X(87-90)
                                                                                       |87-----j11=4---->86:Sub2X(81-86)
                                                                                       \90-----j12=5---->91:Sub2X(91-92)

Added some facility to index SymDofs by their jump name.

Also added some accessors missing in PackerTask, not currently used for anything.
Also changes to my pilot app crap.
Also some warning fixes, though not nearly enough.

If I were a better person, I would separate these changes.

TEST CHANGES:

Added two entries to the integration test ignore_list so that test runs in different directories will not produce diffs (I'm lookin' at you erraser_analytical_closure and swa_rna_loop_sampler!)

cosmetic: beta_strand_homodimer
cosmetic: density_refine_symm
cosmetic: fold_and_dock
cosmetic: symmetric_docking
cosmetic: symmetric_metal_interface_twozinc_design
ignore_list change: fixbb
ignore_list change: erraser_analytical_closure
ignore_list change: swa_rna_loop_sampler

cosmetic changes in log file, mostly new symmetric foldtree output
beta_strand_homodimer
density_refine_symm
fold_and_dock
symmetric_docking
symmetric_metal_interface_twozinc_design

update ignore_list:
in fixbb ignore 'Creating tricubic splines'
in erraser_analytical_closure swa_rna_loop_sampler ignore curent_directory=...



git-svn-id: https://svn.rosettacommons.org/source/trunk/rosetta@48799 6189c7af-7406-0410-a2ce-f22f51069c23

Old repository SHA1: 2507b849db6dbb1fe7bdc0f94eafffa32982b343

+++ rosetta_source/src/apps/pilot/will/xyzStripeHash_test.cc
@@ -97,12 +97,12 @@ int main(int argc, char *argv[]) {
 		std::cout << xyzhash.grid_size() << std::endl;
 		std::cout << xyzhash.natom() << std::endl;
 		std::cout << xyzhash.xdim() << " " << xyzhash.ydim() << " " << xyzhash.zdim() << std::endl;
-		for(Size j = 0; j < xyzhash.natom(); ++j) {
+		for(Size j = 0; j < (int)xyzhash.natom(); ++j) {
 			std::cout << F(12,7,xyzhash.grid_atoms()[j].x) << " ";
 			std::cout << F(12,7,xyzhash.grid_atoms()[j].y) << " ";
 			std::cout << F(12,7,xyzhash.grid_atoms()[j].z) << std::endl;
 		}
-		for(Size j = 0; j < xyzhash.xdim()*xyzhash.ydim()*xyzhash.zdim(); ++j) {
+		for(Size j = 0; j < (int)xyzhash.xdim()*xyzhash.ydim()*xyzhash.zdim(); ++j) {
 			std::cout << xyzhash.grid_stripe()[j].x << " ";
 			std::cout << xyzhash.grid_stripe()[j].y << std::endl;
 		}
@@ -140,7 +140,7 @@ int main(int argc, char *argv[]) {
 	std::cout << "Natom: " << real_natom << " " << xyzhash.natom() << std::endl;
 	{
 		utility::vector1<Vec> hashpts;
-		for(Size i = 0; i < xyzhash.natom(); ++i) {
+		for(Size i = 0; i < (int)xyzhash.natom(); ++i) {
 			hashpts.push_back( Vec( xyzhash.grid_atoms()[i].x, xyzhash.grid_atoms()[i].y, xyzhash.grid_atoms()[i].z ) );
 		}
 		dump_points_pdb(hashpts,"input_xyzhash.pdb");
@@ -149,7 +149,7 @@ int main(int argc, char *argv[]) {
 	
 	double tifc=0.0,th=0.0,ts=0.0,t=0.0;
 	int tot = 0;
-	for(Size i = 0; i < basic::options::option[basic::options::OptionKeys::out::nstruct](); ++i) {
+	for(Size i = 0; i < (int)basic::options::option[basic::options::OptionKeys::out::nstruct](); ++i) {
 		Vec rv( numeric::random::uniform(),numeric::random::uniform(),numeric::random::uniform() );
 		rv.x() = (mx.x()-mn.x()) * rv.x() + mn.x();
 		rv.y() = (mx.y()-mn.y()) * rv.y() + mn.y();
@@ -165,7 +165,7 @@ int main(int argc, char *argv[]) {
 	    Vec tmp = rv+xyzhash.translation();
 		t = time_highres();
 		int safe_nbcount = 0.0;
-		for(Size j = 0; j < xyzhash.natom(); ++j) {
+		for(Size j = 0; j < (int)xyzhash.natom(); ++j) {
 			float const & hx = xyzhash.grid_atoms()[j].x;
 			float const & hy = xyzhash.grid_atoms()[j].y;
 			float const & hz = xyzhash.grid_atoms()[j].z;
