+++ rosetta_source/src/protocols/flexpep_docking/FlexPepDockingAbInitio.cc
@@ -379,7 +379,7 @@ FlexPepDockingAbInitio::convertPSERtoASP(core::pose::Pose& pose, std::set<int>& 
 	using namespace core::chemical;
 	using namespace core::conformation;
 	ResidueTypeSetCAP rsd_set( ChemicalManager::get_instance()->residue_type_set( FA_STANDARD ) );
-	for(int resid = 1; resid <= pose.total_residue(); resid++)
+	for(int resid = 1; resid <= (int)pose.total_residue(); resid++)
 	  {
 	    if(pose.residue_type(resid).has_variant_type(core::chemical::PHOSPHORYLATION) &&
 	       pose.residue_type(resid).name3() == "SER")
@@ -400,7 +400,7 @@ FlexPepDockingAbInitio::convertPSERtoGLU(core::pose::Pose& pose, std::set<int>& 
 	using namespace core::chemical;
 	using namespace core::conformation;
 	ResidueTypeSetCAP rsd_set( ChemicalManager::get_instance()->residue_type_set( FA_STANDARD ) );
-	for(int resid = 1; resid <= pose.total_residue(); resid++)
+	for(int resid = 1; resid <= (int)pose.total_residue(); resid++)
 	  {
 	    if(pose.residue_type(resid).has_variant_type(core::chemical::PHOSPHORYLATION) &&
 	       pose.residue_type(resid).name3() == "SER")
